monorepo: move imported sidecar to 'sidecar'
Move requisite imported controller code to './sidecar'. Modify imports
to allow building. Docker and Makefile also needed some adjustments.
