+++ sidecar/Dockerfile
@@ -1,18 +1,26 @@
+# NOTE: context is expected to be repo root, not sidecar/ dir
+
 #
 # BUILDER
 #
-FROM docker.io/library/golang:1.22.3 AS builder
+FROM docker.io/library/golang:1.22.5 AS builder
 
 WORKDIR /buildroot
 
 # Cache deps before building and copying source, so that we don't need to re-download
 # as much and so that source changes don't invalidate our downloaded layer.
-COPY go.mod go.mod
-COPY go.sum go.sum
+COPY client/ client/
+COPY internal/ internal/
+COPY proto/ proto/
+COPY sidecar/go.mod sidecar/go.mod
+COPY sidecar/go.sum sidecar/go.sum
+
+WORKDIR /buildroot/sidecar
+
 RUN go mod download
 
-COPY cmd/ cmd/
-COPY pkg/ pkg/
+COPY sidecar/cmd/ cmd/
+COPY sidecar/pkg/ pkg/
 
 ENV CGO_ENABLED=0
 
@@ -29,8 +37,8 @@ LABEL description="COSI Storage Sidecar"
 
 LABEL org.opencontainers.image.title="COSI Storage Sidecar"
 LABEL org.opencontainers.image.description="Container Object Storage Interface (COSI) Storage Provisioner Sidecar"
-LABEL org.opencontainers.image.source="https://github.com/kubernetes-sigs/container-object-storage-interface-provisioner-sidecar"
+LABEL org.opencontainers.image.source="https://github.com/kubernetes-sigs/container-object-storage-interface-provisioner-api/sidecar"
 LABEL org.opencontainers.image.licenses="APACHE-2.0"
 
-COPY --from=builder /buildroot/artifacts/objectstorage-sidecar .
+COPY --from=builder /buildroot/sidecar/artifacts/objectstorage-sidecar .
 ENTRYPOINT ["/objectstorage-sidecar"]
