+++ sidecar/cmd/objectstorage-sidecar/main.go
