+++ sidecar/pkg/bucket/bucket_controller.go
@@ -29,12 +29,11 @@ import (
 	kube "k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/klog/v2"
-	"sigs.k8s.io/container-object-storage-interface-api/apis/objectstorage/v1alpha1"
+	"sigs.k8s.io/container-object-storage-interface-api/client/apis/objectstorage/v1alpha1"
 	buckets "sigs.k8s.io/container-object-storage-interface-api/client/clientset/versioned"
 	bucketapi "sigs.k8s.io/container-object-storage-interface-api/client/clientset/versioned/typed/objectstorage/v1alpha1"
-	"sigs.k8s.io/container-object-storage-interface-api/controller/events"
-	"sigs.k8s.io/container-object-storage-interface-provisioner-sidecar/pkg/consts"
-	cosi "sigs.k8s.io/container-object-storage-interface-spec"
+	cosi "sigs.k8s.io/container-object-storage-interface-api/proto"
+	"sigs.k8s.io/container-object-storage-interface-api/sidecar/pkg/consts"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
@@ -73,7 +72,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 		"name", bucket.ObjectMeta.Name)
 
 	if bucket.Spec.BucketClassName == "" {
-		return b.recordError(inputBucket, v1.EventTypeWarning, events.FailedCreateBucket, fmt.Errorf("%w for Bucket %v", consts.ErrUndefinedBucketClassName, bucket.Name))
+		return b.recordError(inputBucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, fmt.Errorf("%w for Bucket %v", consts.ErrUndefinedBucketClassName, bucket.Name))
 	}
 
 	if !strings.EqualFold(bucket.Spec.DriverName, b.driverName) {
@@ -102,12 +101,12 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 		if bucket.Spec.Parameters == nil {
 			bucketClass, err := b.bucketClasses().Get(ctx, bucket.Spec.BucketClassName, metav1.GetOptions{})
 			if kubeerrors.IsNotFound(err) {
-				return b.recordError(inputBucket, v1.EventTypeWarning, events.FailedCreateBucket, err)
+				return b.recordError(inputBucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, err)
 			} else if err != nil {
 				klog.V(3).ErrorS(err, "Error fetching bucketClass",
 					"bucketClass", bucket.Spec.BucketClassName,
 					"bucket", bucket.ObjectMeta.Name)
-				return b.recordError(inputBucket, v1.EventTypeWarning, events.FailedCreateBucket, err)
+				return b.recordError(inputBucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, err)
 			}
 
 			if bucketClass.Parameters != nil {
@@ -128,7 +127,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 		rsp, err := b.provisionerClient.DriverCreateBucket(ctx, req)
 		if err != nil {
 			if status.Code(err) != codes.AlreadyExists {
-				return b.recordError(inputBucket, v1.EventTypeWarning, events.FailedCreateBucket, fmt.Errorf("failed to create bucket: %w", err))
+				return b.recordError(inputBucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, fmt.Errorf("failed to create bucket: %w", err))
 			}
 		}
 
@@ -156,7 +155,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 				klog.V(3).ErrorS(err, "Failed to get bucketClaim",
 					"bucketClaim", ref.Name,
 					"bucket", bucket.ObjectMeta.Name)
-				return b.recordError(bucket, v1.EventTypeWarning, events.FailedCreateBucket, err)
+				return b.recordError(bucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, err)
 			}
 
 			bucketClaim.Status.BucketReady = true
@@ -164,7 +163,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 				klog.V(3).ErrorS(err, "Failed to update bucketClaim",
 					"bucketClaim", ref.Name,
 					"bucket", bucket.ObjectMeta.Name)
-				return b.recordError(bucket, v1.EventTypeWarning, events.FailedCreateBucket, err)
+				return b.recordError(bucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket, err)
 			}
 
 			klog.V(5).Infof("Successfully updated status of bucketClaim: %s, bucket: %s", bucketClaim.ObjectMeta.Name, bucket.ObjectMeta.Name)
@@ -174,7 +173,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 	controllerutil.AddFinalizer(bucket, consts.BucketFinalizer)
 	if bucket, err = b.buckets().Update(ctx, bucket, metav1.UpdateOptions{}); err != nil {
 		klog.V(3).ErrorS(err, "Failed to update bucket finalizers", "bucket", bucket.ObjectMeta.Name)
-		return b.recordError(bucket, v1.EventTypeWarning, events.FailedCreateBucket,
+		return b.recordError(bucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket,
 			fmt.Errorf("failed to update bucket finalizers: %w", err))
 	}
 
@@ -188,7 +187,7 @@ func (b *BucketListener) Add(ctx context.Context, inputBucket *v1alpha1.Bucket) 
 	if _, err = b.buckets().UpdateStatus(ctx, bucket, metav1.UpdateOptions{}); err != nil {
 		klog.V(3).ErrorS(err, "Failed to update bucket status",
 			"bucket", bucket.ObjectMeta.Name)
-		return b.recordError(bucket, v1.EventTypeWarning, events.FailedCreateBucket,
+		return b.recordError(bucket, v1.EventTypeWarning, v1alpha1.FailedCreateBucket,
 			fmt.Errorf("failed to update bucket status: %w", err))
 	}
 
@@ -244,7 +243,7 @@ func (b *BucketListener) Update(ctx context.Context, old, new *v1alpha1.Bucket) 
 		if controllerutil.ContainsFinalizer(bucket, consts.BucketFinalizer) {
 			err = b.deleteBucketOp(ctx, bucket)
 			if err != nil {
-				return b.recordError(bucket, v1.EventTypeWarning, events.FailedDeleteBucket, err)
+				return b.recordError(bucket, v1.EventTypeWarning, v1alpha1.FailedDeleteBucket, err)
 			}
 
 			controllerutil.RemoveFinalizer(bucket, consts.BucketFinalizer)
