+++ sidecar/pkg/provisioner/server.go
@@ -21,9 +21,8 @@ import (
 
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
-
 	"k8s.io/klog/v2"
-	cosi "sigs.k8s.io/container-object-storage-interface-spec"
+	cosi "sigs.k8s.io/container-object-storage-interface-api/proto"
 )
 
 type COSIProvisionerServer struct {
